+++ providers/darwin/process_darwin_test.go
@@ -28,8 +28,10 @@ import (
 	"github.com/elastic/go-sysinfo/internal/registry"
 )
 
-var _ registry.HostProvider = darwinSystem{}
-var _ registry.ProcessProvider = darwinSystem{}
+var (
+	_ registry.HostProvider    = darwinSystem{}
+	_ registry.ProcessProvider = darwinSystem{}
+)
 
 func TestKernProcInfo(t *testing.T) {
 	var p process